+++ src/resources/views/components/help-tooltip.blade.php
@@ -1,0 +1,2 @@
+&nbsp;
+<a href="#" data-bs-toggle="tooltip" data-bs-placement="right" data-bs-title="{{ $slot }}"><small>{!! icon('help', 'info') !!}</small></a>