+++ src/resources/views/product/_form_seo.blade.php
@@ -1,13 +1,16 @@
-<div class="mb-3">
-    {{ Form::text('ext_title', null, [
-            'class' => 'form-control form-control-sm' . ($errors->has('ext_title') ? ' is-invalid': ''),
-            'placeholder' => __('Extended/SEO Title')
-        ])
-    }}
-    @if ($errors->has('ext_title'))
-        <div class="invalid-feedback">{{ $errors->first('ext_title') }}</div>
-    @endif
-</div>
+@if (request()->routeIs('vanilo.admin.product.edit') || request()->routeIs('vanilo.admin.product.create'))
+    <div class="mb-3">
+        {{ Form::text('gtin', null, [
+                'class' => 'form-control form-control-sm' . ($errors->has('gtin') ? ' is-invalid' : ''),
+                'placeholder' => __('Global Trade Identification Number'),
+                'maxlength' => 255,
+            ])
+        }}
+        @if ($errors->has('gtin'))
+            <div class="invalid-feedback">{{ $errors->first('gtin') }}</div>
+        @endif
+    </div>
+@endif
 
 <div class="mb-3">
     {{ Form::textarea('meta_description', null, [
@@ -32,16 +35,3 @@
     @endif
 </div>
 
-@if (request()->routeIs('vanilo.admin.product.edit') || request()->routeIs('vanilo.admin.product.create'))
-    <div class="mb-3">
-        {{ Form::text('gtin', null, [
-                'class' => 'form-control form-control-sm' . ($errors->has('gtin') ? ' is-invalid' : ''),
-                'placeholder' => __('Global Trade Identification Number'),
-                'maxlength' => 255,
-            ])
-        }}
-        @if ($errors->has('gtin'))
-            <div class="invalid-feedback">{{ $errors->first('gtin') }}</div>
-        @endif
-    </div>
-@endif